requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.5
django-anymail==2.2
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==2.0.7
django-filter==1.1.0
django-guardian==1.4.9
django-rest-swagger==2.2.0
django-storages==1.6.6
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.4
requests==2.18.4
simplejson==3.14.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
